Disconnect TextRegion to PostContainer context
[#140605547](https://www.pivotaltracker.com/story/show/140605547)
