+++ src/lib/jello.js
@@ -1,5 +1,14 @@
 import TWEEN from 'tween.js'
 
+// -------------------------------------
+
+export const isLink = (target) => {
+  if (target.nodeName.toLowerCase() === 'a') { return true }
+  const parent = target.closest ? target.closest('a') : target.parentNode
+  return parent && parent.nodeName.toLowerCase() === 'a'
+}
+
+// -------------------------------------
 let memoizedIsAndroid
 let memoizedIsChrome
 let memoizedIsElloAndroid
